Include a sphinx-based documentation
This includes documentation following the standard Python style. Sphinx
can be used to build this into HTML, and it should be trivial to hook
this up to something like RTD to publish them too.

Locally, `tox -e docs` can be used to builds docs. It may make sense to
include this in CI too to make sure docs continue to build.