site_cache_dir: Use /var/cache again instead of /var/tmp on UNIX
This directory was changed from /var/cache to /var/tmp in #148 due to
permissions issues. However, /var/tmp is an insecure location to store
anything with a predictable filename, because any other user could
have written it first. This leads to vulnerabilities categorized under
CWE-377 and CAPEC-149.

To deal with the permissions issues, applications should put their own
cache data in a subdirectory of /var/cache (e.g. /var/cache/cups), and
the application’s package is responsible for ensuring the subdirectory
exists and giving it the correct permissions.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>