Simplify logic for choosing `get_win_folder`’s value
Before this change, `windows.py` had multiple different `import ctypes`
and `import winreg` lines. `windows.py` did this for a specific reason.
The `get_win_folder_via_ctypes()` function depends on the `ctypes`
module, and the `get_win_folder_from_registry()` functions depends on
the `winreg` modules, but `windows.py` as a whole is supposed to
continue to work even if `ctypes` and `winreg` are not available.

Before this change, `windows.py` would try to avoid errors by only
calling either of those two functions if the module that that function
requires is not available. This meant that there had to be two
`import ctypes` lines and two `import winreg` lines. The first one would
be inside the body of the function that that depends on the module, and
the second one would be in a test that determines if the function should
be called. The idea was to only call `get_win_folder_via_ctypes()` and
`get_win_folder_from_registry()` if their respective modules are
available.

This change simplifies `windows.py`. Previously, those two functions
were always defined but not always called. Now, those two fuctions are
not always defined and not always called. This means that there’s now
only one `import ctypes` line and only one `import winreg` line. As an
added bonus, those two `import` statements are now at the top of the
file (like the rest of the `import` statements) and the
`_pick_get_win_folder()` function is now simpler.

The main motivation behind this change is to make it easier to create a
future change. That future change will add additional code to
`windows.py` that depends on the `ctypes` module. This change will allow
me to add that additional code without adding an additional
`import ctypes` line.